backend_ai/requirements.txt
langchain
langchain-core
langchain-google-genai
google-generativeai
transformers
numpy
scikit-learn
python-dotenv
ipykernel
langchainhub
yfinance
plotly
firebase
firebase_admin
fastapi
uvicorn
bcrypt
passlib
pyjwt